+++ systole/detectors/__init__.py
@@ -1,7 +1,9 @@
+from .christov import christov
 from .hamilton import hamilton
 from .pan_tompkins import pan_tompkins
 
 __all__ = [
     "pan_tompkins",
     "hamilton",
+    "christov",
 ]